skip secret deletion

+++ test/functional/corerp/corerptest.go
@@ -13,6 +13,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/client-go/discovery"
@@ -59,7 +60,8 @@ type CoreRPTest struct {
 	PostDeleteVerify func(ctx context.Context, t *testing.T, ct CoreRPTest)
 
 	// Object Name => map of secret keys and values
-	Secrets map[string]map[string]string
+	Secrets            map[string]map[string]string
+	SkipSecretDeletion bool
 }
 
 type TestOptions struct {
@@ -73,12 +75,13 @@ func NewTestOptions(t *testing.T) TestOptions {
 
 func NewCoreRPTest(t *testing.T, name string, steps []TestStep, secrets map[string]map[string]string, initialResources ...unstructured.Unstructured) CoreRPTest {
 	return CoreRPTest{
-		Options:          NewCoreRPTestOptions(t),
-		Name:             name,
-		Description:      name,
-		Steps:            steps,
-		Secrets:          secrets,
-		InitialResources: initialResources,
+		Options:            NewCoreRPTestOptions(t),
+		Name:               name,
+		Description:        name,
+		Steps:              steps,
+		Secrets:            secrets,
+		InitialResources:   initialResources,
+		SkipSecretDeletion: false,
 	}
 }
 
@@ -130,7 +133,7 @@ func (ct CoreRPTest) CreateSecrets(ctx context.Context) error {
 					Data: data,
 				}, v1.CreateOptions{})
 			if err != nil {
-				return fmt.Errorf("failed to create secret %s", err.Error())
+				return err
 			}
 		}
 	}
@@ -208,19 +211,22 @@ func (ct CoreRPTest) Test(t *testing.T) {
 	})
 
 	t.Logf("Creating secrets if provided")
-	_ = ct.CreateSecrets(ctx)
-	/*
-		if err != nil {
+	err := ct.CreateSecrets(ctx)
+	if err != nil {
+		if errors.IsAlreadyExists(err) {
+			// Do not stop the test if the same secret exists
+			t.Logf("the secret already exists %v", err)
+		} else {
 			t.Errorf("failed to create secrets %v", err)
 		}
-	*/
+	}
 
 	// Inside the integration test code we rely on the context for timeout/cancellation functionality.
 	// We expect the caller to wire this out to the test timeout system, or a stricter timeout if desired.
 
 	require.GreaterOrEqual(t, len(ct.Steps), 1, "at least one step is required")
 	defer ct.CleanUpExtensionResources(ct.InitialResources)
-	err := ct.CreateInitialResources(ctx)
+	err = ct.CreateInitialResources(ctx)
 	require.NoError(t, err, "failed to create initial resources")
 
 	success := true
@@ -315,13 +321,15 @@ func (ct CoreRPTest) Test(t *testing.T) {
 			}
 		}
 	}
-	/*
+
+	if !ct.SkipSecretDeletion {
 		t.Logf("Deleting secrets")
 		err = ct.DeleteSecrets(ctx)
 		if err != nil {
 			t.Errorf("failed to delete secrets %v", err)
 		}
-	*/
+	}
+
 	// Custom verification is expected to use `t` to trigger its own assertions
 	if ct.PostDeleteVerify != nil {
 		t.Logf("running post-delete verification for %s", ct.Description)
